requirements.txt
fastapi
uvicorn[standard]
pillow
opencv-python-headless
numpy
python-multipart